Add SNODAS backfill feature #7

+++ snowmelt/__init__.py
@@ -30,7 +30,19 @@ def print_dashes(length=64):
     print '-' * length
 
 
-def prepare_source_data_for_date(process_date, src_dir, save_tiff=True):
+def get_src_dir_by_date(proc_date):
+    # Helper function to find the source data, which is stored in
+    # a few different places on rsgis-base.
+    if proc_date.year > 2012:
+        return config.SRC_DIR
+    elif proc_date.year == 2012:
+        return config.ARCHIVE_DIR_2012
+    else:
+        month_path = proc_date.strftime('%Y/%B')
+        return os.path.join(config.ARCHIVE_DIR, month_path)
+
+
+def prepare_source_data_for_date(process_date, src_dir, conus_tiff_only=True):
     ''' Builds an unzip directory and extracts data from source files
     for a given day. 
     Returns the directory path to the unzipped files,
@@ -66,15 +78,23 @@ def prepare_source_data_for_date(process_date, src_dir, save_tiff=True):
         print_dashes()
         return None
 
-    if save_tiff:
-        mkdir_p(us_tif_dir)
+    mkdir_p(us_tif_dir)
+    mkdir_p(unzip_dir)
 
     # Loop through our filenames and do the unzipping and other set up.
-    mkdir_p(unzip_dir)
     for filename in snodas_src_files:
+        
+        # Cobble together all the file names.
         src_file = os.path.join(src_dir, filename)
         unzip_file = os.path.join(unzip_dir, filename)
         ready_file = unzip_file + '.bil'
+        shgtif = os.path.join(us_tif_dir, filename + 'alb.tif')
+
+        # Don't bother doing anything if tiff already exists in conus_tiff_only mode.
+        if conus_tiff_only and os.path.isfile(shgtif):
+            print 'CONUS SHG tiff already exists:', shgtif
+            continue
+
         if not os.path.isfile(ready_file):
             print 'Processing source to output file:', ready_file
             UnzipLinux(src_file, unzip_file)
@@ -83,14 +103,12 @@ def prepare_source_data_for_date(process_date, src_dir, save_tiff=True):
             print 'Using existing source file:', ready_file 
         
         # Save a full version of the day's data set.
-        shgtif = os.path.join(us_tif_dir, filename + 'alb.tif')
-        if save_tiff:
-            if not os.path.isfile(shgtif):
-                print 'Saving CONUS SHG tiff file:', shgtif
-                ReprojUseWarpBil(ready_file, shgtif, nodata=nodata_val,
-                                 tr_x='1000', tr_y='-1000')
-            else:
-                print 'CONUS SHG tiff already exists:', shgtif
+        if not os.path.isfile(shgtif):
+            print 'Saving CONUS SHG tiff file:', shgtif
+            ReprojUseWarpBil(ready_file, shgtif, nodata=nodata_val,
+                             tr_x='1000', tr_y='-1000')
+        else:
+            print 'CONUS SHG tiff already exists:', shgtif
 
     print_dashes()
     return unzip_dir
@@ -396,7 +414,8 @@ def min_box_os(ext1, ext2, cellsize):
 
 def RawFileManip(file_noext, masterhdr):
     ''' Replaces header with custom header file and renames .dat to .bil '''
-    os.remove(file_noext + ".Hdr")
+    if os.path.exists(file_noext + ".Hdr"):
+        os.remove(file_noext + ".Hdr")
     shutil.copy(masterhdr, file_noext + ".hdr")
     if os.path.exists(file_noext + ".bil"):
         os.remove(file_noext + ".bil")
@@ -596,11 +615,21 @@ def WriteToDSS(inasc, outdss, dtype, path, dunits='MM'):
     bname = os.path.basename(inasc)
     os.chdir(pname)
     
-    asc2dsscmd = os.path.join(config.TOP_DIR, 'Asc2DssGridUtility', 
-                              'asc2dssGriddash')
+    # Old subroutine
+    # asc2dsscmd = os.path.join(config.TOP_DIR, 'Asc2DssGridUtility', 
+    #                           'asc2dssGriddash')
+    # cmdlist = [
+    #     'python', asc2dsscmd, 'gridtype=SHG', 'dunits=' + dunits,
+    #     'dtype=' + dtype, 'in=' + bname, 'dss=' + outdss, 'path=' + path
+    # ]
+    asc2dsscmd = os.path.join(
+        config.TOP_DIR, 'Asc2DssGridUtility', 'asc2dssBash'
+    )
+    quote_str = '"{0}"'
     cmdlist = [
-        'python', asc2dsscmd, 'gridtype=SHG', 'dunits=' + dunits,
-        'dtype=' + dtype, 'in=' + bname, 'dss=' + outdss, 'path=' + path
+        asc2dsscmd, 'gridtype=SHG', 'dunits=' + quote_str.format(dunits),
+        'dtype=' + quote_str.format(dtype), 'in=' + quote_str.format(bname), 
+        'dss=' + quote_str.format(outdss), 'path=' + quote_str.format(path)
     ]
     if not config.SUBPROCESS_QUIET:
         print pname
